+++ src/cfnlint/rules/functions/GetAz.py
@@ -37,7 +37,7 @@ class GetAz(CloudFormationLintRule):
             getaz_value = getaz_obj[-1]
             if isinstance(getaz_value, six.string_types):
                 if getaz_value != '' and getaz_value not in cfn.regions:
-                    message = "GetAZs should be of empty or string of valid region for {0}"
+                    message = 'GetAZs should be of empty or string of valid region for {0}'
                     matches.append(RuleMatch(
                         getaz_obj[:-1], message.format('/'.join(map(str, getaz_obj[:-1])))))
             elif isinstance(getaz_value, dict):
@@ -45,15 +45,15 @@ class GetAz(CloudFormationLintRule):
                     if isinstance(getaz_value, dict):
                         for key, value in getaz_value.items():
                             if key != 'Ref' or value != 'AWS::Region':
-                                message = "GetAZs should be of Ref to AWS::Region for {0}"
+                                message = 'GetAZs should be of Ref to AWS::Region for {0}'
                                 matches.append(RuleMatch(
                                     getaz_obj[:-1], message.format('/'.join(map(str, getaz_obj[:-1])))))
                     else:
-                        message = "GetAZs should be of Ref to AWS::Region for {0}"
+                        message = 'GetAZs should be of Ref to AWS::Region for {0}'
                         matches.append(RuleMatch(
                             getaz_obj[:-1], message.format('/'.join(map(str, getaz_obj[:-1])))))
                 else:
-                    message = "GetAZs should be of Ref to AWS::Region for {0}"
+                    message = 'GetAZs should be of Ref to AWS::Region for {0}'
                     matches.append(RuleMatch(
                         getaz_obj[:-1], message.format('/'.join(map(str, getaz_obj[:-1])))))
         return matches