+++ src/cfnlint/rules/resources/ectwo/SecurityGroupIngress.py
@@ -31,7 +31,7 @@ class SecurityGroupIngress(CloudFormationLintRule):
         """Check VPC Values"""
         matches = list()
         if not value.startswith('sg-'):
-            message = "Security Group Id must have a valid Identifier {0}"
+            message = 'Security Group Id must have a valid Identifier {0}'
             matches.append(
                 RuleMatch(path, message.format('/'.join(map(str, path)))))
         return matches
@@ -50,7 +50,7 @@ class SecurityGroupIngress(CloudFormationLintRule):
             parameter_type = parameter_properties.get('Type')
             if parameter_type not in allowed_types:
                 path_error = ['Parameters', value, 'Type']
-                message = "Security Group Id Parameter should be of type [{0}] for {1}"
+                message = 'Security Group Id Parameter should be of type [{0}] for {1}'
                 matches.append(
                     RuleMatch(
                         path_error,
@@ -59,16 +59,16 @@ class SecurityGroupIngress(CloudFormationLintRule):
                             '/'.join(map(str, path_error)))))
         if value in resources:
             resource = resources.get(value, {})
-            resource_type = resource.get('Type', "")
+            resource_type = resource.get('Type', '')
             if resource_type != 'AWS::EC2::SecurityGroup':
-                message = "Security Group Id resources should be of type AWS::EC2::SecurityGroup for {0}"
+                message = 'Security Group Id resources should be of type AWS::EC2::SecurityGroup for {0}'
                 matches.append(
                     RuleMatch(path, message.format('/'.join(map(str, path)))))
             else:
                 resource_properties = resource.get('Properties', {})
                 vpc_property = resource_properties.get('VpcId', None)
                 if not vpc_property:
-                    message = "Security Group Id should reference a VPC based AWS::EC2::SecurityGroup for {0}"
+                    message = 'Security Group Id should reference a VPC based AWS::EC2::SecurityGroup for {0}'
                     matches.append(
                         RuleMatch(path, message.format('/'.join(map(str, path)))))
 
@@ -79,7 +79,7 @@ class SecurityGroupIngress(CloudFormationLintRule):
         """Automatic failure for certain functions"""
 
         matches = list()
-        message = "Use Ref, FindInMap, or string values for {0}"
+        message = 'Use Ref, FindInMap, or string values for {0}'
         matches.append(
             RuleMatch(path, message.format('/'.join(map(str, path)))))
         return matches
@@ -93,8 +93,8 @@ class SecurityGroupIngress(CloudFormationLintRule):
             # Check that SourceSecurityGroupName isn't specified
             if properties.get('SourceSecurityGroupName', None):
                 path_error = path[:] + ['SourceSecurityGroupName']
-                message = "SourceSecurityGroupName shouldn't be specified for " \
-                          "Vpc Security Group at {0}"
+                message = 'SourceSecurityGroupName shouldn\'t be specified for ' \
+                          'Vpc Security Group at {0}'
                 matches.append(
                     RuleMatch(path_error, message.format('/'.join(map(str, path_error)))))
 
@@ -112,8 +112,8 @@ class SecurityGroupIngress(CloudFormationLintRule):
 
             if properties.get('SourceSecurityGroupId', None):
                 path_error = path[:] + ['SourceSecurityGroupId']
-                message = "SourceSecurityGroupId shouldn't be specified for " \
-                          "Non-Vpc Security Group at {0}"
+                message = 'SourceSecurityGroupId shouldn\'t be specified for ' \
+                          'Non-Vpc Security Group at {0}'
                 matches.append(
                     RuleMatch(path_error, message.format('/'.join(map(str, path_error)))))