+++ src/cfnlint/rules/templates/Base.py
@@ -52,7 +52,7 @@ class Base(CloudFormationLintRule):
         for x in cfn.template:
             top_level.append(x)
             if x not in self.valid_keys:
-                message = "Top level item {0} isn't valid"
+                message = 'Top level item {0} isn\'t valid'
                 matches.append(RuleMatch(
                     [x],
                     message.format(x)
@@ -60,7 +60,7 @@ class Base(CloudFormationLintRule):
 
         for y in self.required_keys:
             if y not in top_level:
-                message = "Missing top level item {0} to file module"
+                message = 'Missing top level item {0} to file module'
                 matches.append(RuleMatch(
                     ['AWSTemplateFormatVersion'],
                     message.format(y)